requirements.txt
# ============================================================================
# SMS SPAM DETECTION PROJECT - REQUIREMENTS
# ============================================================================
# Installation: pip install -r requirements.txt

# ----------------------------------------------------------------------------
# CORE LIBRARIES
# ----------------------------------------------------------------------------
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# ----------------------------------------------------------------------------
# DEEP LEARNING FRAMEWORKS
# ----------------------------------------------------------------------------
# TensorFlow/Keras (pour modèles classiques)
tensorflow==2.13.0
keras==2.13.1

# PyTorch (pour modèles avancés et meta-learning)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# ----------------------------------------------------------------------------
# NLP & TRANSFORMERS
# ----------------------------------------------------------------------------
# HuggingFace pour BERT, DistilBERT, etc.
transformers==4.31.0
tokenizers==0.13.3
datasets==2.14.4

# Sentence Transformers (pour embeddings et RAG)
sentence-transformers==2.2.2

# NLTK (prétraitement classique)
nltk==3.8.1

# SpaCy (alternative preprocessing)
spacy==3.6.1

# TextBlob (sentiment analysis - pour multi-task)
textblob==0.17.1


faiss-cpu==1.7.4


#


langchain==0.0.277
langchain-community==0.0.1


pinecone-client==2.2.4


pytorch-lightning==2.0.6


learn2learn==0.1.7

higher==0.2.1


plotly==5.16.1

# WordCloud
wordcloud==1.9.2

# SHAP (explainability)
shap==0.42.1

# LIME (explainability)
lime==0.2.0.1

# Yellowbrick (ML visualization)
yellowbrick==1.5


streamlit==1.26.0

# Gradio (alternative pour UI)
gradio==3.41.2

flask==2.3.3
flask-cors==4.0.0

fastapi==0.103.0
uvicorn==0.23.2


jupyter==1.0.0
notebook==7.0.2
ipywidgets==8.1.0


tqdm==4.66.1


pyyaml==6.0.1
python-dotenv==1.0.0

optuna==3.3.0


wandb==0.15.8

imbalanced-learn==0.11.0

# Text preprocessing
ftfy==6.1.1
contractions==0.1.73
unidecode==1.3.6

# Regular expressions
regex==2023.8.8


joblib==1.3.2

# Memory profiling
memory-profiler==0.61.0

# GPU monitoring
gpustat==1.1

# ----------------------------------------------------------------------------
# TESTING & QUALITY
# ----------------------------------------------------------------------------
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0

# ----------------------------------------------------------------------------
# ADDITIONAL MODELS & TOOLS
# ----------------------------------------------------------------------------
# XGBoost (si comparaison avec ML)
xgboost==1.7.6

# CatBoost (alternative)
catboost==1.2

# Gensim (Word2Vec, Doc2Vec)
gensim==4.3.1


protobuf==3.20.3
typing-extensions==4.7.1
packaging==23.1

